+++ server/src/config/runtime.py
@@ -87,15 +87,15 @@ class RuntimeConfig:
 
     def reload_backup_assignments(self, _=None):
         for service_id, service in self.context.app.services.items():
-            err_message = f" For 'backup security' purposes all backup configs must have an entry in the runtime config (even if the service is not assigned to a server)"
+            err_message = f" For 'backup security' purposes all backup configs should have an entry in the runtime config (even if the service is not assigned to a server)"
             if not service_id in self.assignments.keys() and len(service.backup_configs) > 0:
-                raise GenericConfigException(f"Service {service_id} has backup configs, but is not in the runtime config. {err_message}")
+                logger.error(f"Service {service_id} has backup configs, but is not in the runtime config. {err_message}")
             backup_assignment_service_config = self.backup_assignments.get(service_id, {})
             for backup_config in service.backup_configs:
                 if not backup_config.id_str in backup_assignment_service_config.keys():
-                    raise GenericConfigException(f"Backup config {backup_config.id_str} for service {service_id} is not in the runtime config. {err_message}")
+                    logger.error(f"Backup config {backup_config.id_str} for service {service_id} is not in the runtime config. {err_message}")
                 backup_assignments: list[dict] = backup_assignment_service_config.get(backup_config.id_str, [])
                 for backup_assignment in backup_assignments:
                     if backup_assignment.get("server") is None or backup_assignment.get("storage") is None:
-                        raise GenericConfigException("Backup assignment is missing required fields: 'server' and 'storage'")
+                        logger.error("Backup assignment is missing required fields: 'server' and 'storage'")
                 backup_config.targets = backup_assignments